refactor: Split code to improve readability
